+++ components/Main/location/locationSelectBox.tsx
@@ -11,9 +11,9 @@ import {
 } from '@chakra-ui/react';
 import { BiCurrentLocation } from 'react-icons/bi';
 
-import { SEOUL_AREA } from './constants';
+import { SEOUL_AREA } from '../constants';
+import { ILocationSelectBox } from '../types';
 import LocationListItem from './locationListItem';
-import { ILocationSelectBox } from './types';
 
 export default function LocationSelectBox({
   location,
