+++ pages/main.tsx
@@ -1,4 +1,4 @@
-import CakeMain from '@/components/Main/cakeMain';
+import CakeMain from '@/components/Main/cake/cakeMain';
 
 export default function Main() {
   return <CakeMain />;
